Fixed a double free crash.
Fixed a crash that happend in some cases when compressobj destructor free's outstream's data. It checks if ouststream's data is not null before free'ing but data is never set to null after free'ing. Then it also tries to free outstream's data on exit and this leads to us trying to free the data twice. This leads to the following error when using pylzma: python: double free or corruption (out): 0x000000000166d5a0 ***